[REFACTOR]: AddUser 리펙토링
